+++ spec/scrap_kit/recipe_spec.rb
@@ -19,9 +19,35 @@ RSpec.describe ScrapKit::Recipe do
 
     expect(output).to eq(
       posts: [
+        { title: "Usando OpenStruct" },
+        { title: "Aprendiendo a usar arrays en JavaScript" },
         { title: "APIs de Internacionalización en JavaScript" },
         { title: "Ejecutando comandos desde Ruby" },
         { title: "Usando Higher-Order Components" }
+      ]
+    )
+  end
+
+  it "Load recipe with selector array" do
+    recipe = ScrapKit::Recipe.load(
+      url: "https://status.heroku.com/",
+      attributes: {
+        results: {
+          selector: [".up-time-chart", { ".region-header .u-margin-Tm": "REGION" }],
+          children_attributes: {
+            region: "h4:first-child",
+            uptime: "h4:last-child"
+          }
+        }
+      }
+    )
+
+    output = recipe.run
+
+    expect(output).to eq(
+      results: [
+        { region: "US REGION", uptime: "99.999709%" },
+        { region: "EU REGION", uptime: "99.999994%" }
       ]
     )
   end