+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.1.2] 2020-06-18
+
+### Changed/Added
+- Add support for selector arrays to filter elements based on conditions
+
+## [0.1.1] 2020-05-18
+
+### Changed/Added
+- Update activesupport requirement from = 6.0.2.1 to = 6.0.3.1
+
 ## [0.1.0] 2020-02-09
 
 ### Changed/Added